Modified some tests (and merged to proper location)
